backend/requirements.txt
flask
flask-cors
pycryptodome
pandas
numpy
scikit-learn
tensorflow
openpyxl
matplotlib
seaborn
